fix. Изменена структура файловой системы для страницы OfferPage
